added options and menu choices for setting the device, version, and MAC address
In this updated code, options for setting the device, version, and MAC address have been added to the menu. When the user selects these options, they will be prompted to enter the corresponding values, and those values will be stored in the device, version, and mac variables, respectively. The set values will be displayed for confirmation.

The report generation (GenerateReport()) function now includes additional checks to ensure that all the required variables (brokerAddress, brokerPort, device, version, and mac) are set before generating the report. If any of these variables are missing, an appropriate error message will be displayed.